+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RedLeftAuto.java
@@ -1,36 +1,40 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
+import com.acmerobotics.roadrunner.geometry.Pose2d;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
 @Autonomous(name="Red Left Auto")
-public class RedLeftAuto extends LinearOpMode{
+public class RedLeftAuto extends LinearOpMode {
+
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
     Servo leftGripper;
     Servo rightGripper;
-    Pose2d currentPose;
-    int purplePos;
+    DcMotor slideMotor1;
+    DcMotor slideMotor2;
+    Pose2d selectPose = new Pose2d(0,0,0);
+    double selectTurn = 0;
+    double currentHeading = 0;
+    double boardOffset = 0;
     boolean armUp = false;
     int slidePos;
     double rampPos = 0.501;
     Servo arm;
     double slidePosError;
     double slidePower;
-    DcMotor slideMotor1;
-    DcMotor slideMotor2;
-    double dist;
+    double waitTime;
+
     @Override
-    public void runOpMode() throws InterruptedException {
+    public void runOpMode() throws InterruptedException
+    {
         leftSensor = hardwareMap.get(DistanceSensor.class, "checkLeft");
         rightSensor = hardwareMap.get(DistanceSensor.class, "checkRight");
 
@@ -55,87 +59,68 @@ public class RedLeftAuto extends LinearOpMode{
         slideMotor2.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
+        Pose2d startPose = new Pose2d(-38.25, -63.75, Math.toRadians(90));
+        drive.setPoseEstimate(startPose);
 
         waitForStart();
 
-        TrajectorySequence approachSpikeMarks = drive.trajectorySequenceBuilder(new Pose2d(-38.5, -63.75, Math.toRadians(-90)))
-                .strafeRight(3)
-                .forward(32.5)
+        TrajectorySequence approachSpikeMarks = drive.trajectorySequenceBuilder(startPose)
+                //Move into spike mark area
+                .splineToLinearHeading(new Pose2d(-35, -30.25, startPose.getHeading()), startPose.getHeading())
+                //Determine where prop is located
+                .addDisplacementMarker(() -> {
+                    if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
+                        selectPose = new Pose2d(-39, -31.25, Math.toRadians(180));
+                        selectTurn = 90;
+                        boardOffset = 5.75;
+                        waitTime = 3.25;
+                    } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
+                        selectPose = new Pose2d(-32, -31.25,0);
+                        selectTurn = -90;
+                        boardOffset = -5.75;
+                        waitTime = 0;
+                    } else {
+                        selectPose = new Pose2d(-35.5, -31.5, Math.toRadians(90));
+                        selectTurn = 0;
+                        boardOffset = 0;
+                        waitTime = 2;
+                    }
+                })
                 .build();
-
         drive.followTrajectorySequence(approachSpikeMarks);
 
-        if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-            purplePos = 1;
-            dist = -4;
-            TrajectorySequence placePurpleLeft = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
-                    .turn(Math.toRadians(90))
-                    .forward(3.5)
-                    .build();
-
-            drive.followTrajectorySequence(placePurpleLeft);
-
-            leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-
-            TrajectorySequence backLeft = drive.trajectorySequenceBuilder(placePurpleLeft.end())
-                    .waitSeconds(1)
-                    .back(3.5)
-                    .strafeLeft(28)
-                    .turn(Math.toRadians(180))
-                    .build();
-
-            drive.followTrajectorySequence(backLeft);
-
-            currentPose = backLeft.end();
-
-        } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-            purplePos = 2;
-            dist = 4;
-            TrajectorySequence placePurpleRight = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
-                    .turn(Math.toRadians(-90))
-                    .forward(3.5)
-                    .build();
-
-            drive.followTrajectorySequence(placePurpleRight);
-
-            leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-
-            TrajectorySequence backRight = drive.trajectorySequenceBuilder(placePurpleRight.end())
-                    .waitSeconds(1)
-                    .back(3.5)
-                    .strafeRight(28)
-                    .build();
-
-            drive.followTrajectorySequence(backRight);
-
-            currentPose = backRight.end();
-
-        } else {
-            purplePos = 3;
-            dist = 0;
-            leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-            TrajectorySequence back = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
-                    .back(28)
-                    .turn(Math.toRadians(-90))
-                    .build();
-
-            drive.followTrajectorySequence(back);
-
-            currentPose = back.end();
-        }
-
-
-        TrajectorySequence driveToBoard = drive.trajectorySequenceBuilder(currentPose)
-                .back(23.5)
-                .strafeLeft(50)
-                .forward(47)
-                .waitSeconds(0.5)
-                .forward(47)
-                .strafeRight(26 + dist)
-                .forward(5)
+        TrajectorySequence placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
+                //Turn to proper spike mark
+                .turn(Math.toRadians(selectTurn))
+                //Drive up to proper spike mark
+                .lineToLinearHeading(selectPose)
+                //Drop purple pixel
+                .addDisplacementMarker(() -> {
+                    leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
+                })
+                //Back away from purple pixel
+                .back(4.5)
+                //Move back to safe turning position
+                .lineToLinearHeading(new Pose2d(-37, -48, selectPose.getHeading()))
+                //Turn towards back wall
+                .turn(Math.toRadians(179.999)-selectPose.getHeading())
+                //Drive towards back wall
+                .lineToLinearHeading(new Pose2d(-58,-48, Math.toRadians(180)))
+                //Turn towards opposing side
+                .turn(Math.toRadians(-90))
+                //Drive towards opposing side
+                .lineToLinearHeading(new Pose2d(-58,-9, Math.toRadians(90)))
+                //Turn towards center of field
+                .turn(Math.toRadians(-90))
+                //Drive towards center of field
+                .lineToLinearHeading(new Pose2d(15,-9,0))
+                //Wait for teammate to finish
+                .waitSeconds(waitTime)
+                //Drive to board
+                .splineToLinearHeading(new Pose2d(39,-32.5 + boardOffset,0),Math.toRadians(-90))
                 .build();
 
-        drive.followTrajectorySequence(driveToBoard);
+        drive.followTrajectorySequence(placePurplePixel);
 
         slidePos = 600;
         armUp = true;
@@ -154,23 +139,20 @@ public class RedLeftAuto extends LinearOpMode{
         slideMotor1.setPower(0);
         slideMotor2.setPower(0);
 
-        TrajectorySequence placeYellow = drive.trajectorySequenceBuilder(driveToBoard.end())
-                .forward(9)
-                .build();
-        drive.followTrajectorySequence(placeYellow);
-        currentPose = placeYellow.end();
-
-
-        rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
-
-        TrajectorySequence backUp = drive.trajectorySequenceBuilder(currentPose)
+        TrajectorySequence dropOnBoard = drive.trajectorySequenceBuilder(placePurplePixel.end())
+                //Drive to place yellow pixel
+                .forward(49.5 - placePurplePixel.end().getX())
+                //Drop yellow pixel
+                .addDisplacementMarker(() -> {
+                    rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
+                })
+                //Wait for gripper to open
                 .waitSeconds(0.4)
+                //Back up to release pixel if trapped against backdrop
                 .back(5)
                 .build();
-
-        drive.followTrajectorySequence(backUp);
+        drive.followTrajectorySequence(dropOnBoard);
     }
-
     public double Ramp(double firstPos, double secondPos, boolean selectPos)
     {
         if(selectPos){
