+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -67,8 +67,8 @@ public class DriveConstants {
      */
     public static double MAX_VEL = 50;
     public static double MAX_ACCEL = 30;
-    public static double MAX_ANG_VEL = 2;
-    public static double MAX_ANG_ACCEL = 1;
+    public static double MAX_ANG_VEL = 3;
+    public static double MAX_ANG_ACCEL = 1.5;
 
     /*
      * Adjust the orientations here to match your robot. See the FTC SDK documentation for details.
